services/web/requirements.txt
dash==2.5.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
openpyxl==3.1.2
pandas==2.2.1
plotly==5.21.0
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.29
Werkzeug==3.0.2
gunicorn==20.1.0